Anki add: move menu items Add/Delete to first, not last, position
This is because it's the next action from the user, to know if the card was found locally or remotely.
In the latter case, the user should decide to Add, or not, the card.
This makes that more visible, by being in line (along the left margin) with the definition.