Anki add: iterate through multiple matches (like for Empties)
This fixes a bug so that it's now possible to also Delete a card in the middle of set of matching cards.
However, the state machine between the different commands is getting more fragile.
Time to zoom out and reconsider this approach ...

Some new coloring logic (WIP), and menu formatting.