Anki add: enable readline completion of deck names when switching decks
This is done by pushing deck names onto the readline history/stack and
then popping them off again after a deck is (not) selected. Also, the
completion of dictionary words is disabled when completing deck names.

Since there seems to be only one readline history, it's a bit of a hack
to share the same stack for dictionary names as well as deck names, but
since the set of deck names is relatively quite small, this might be a
tolerable trade-off.